# apis/calendar/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0-rc.2
  description: |
    离线中国大陆日历资产 API，提供节假日、调休、农历、节气和干支查询。
    本版本是契约整体替换，1.x 在本项目中不承诺兼容。
  x-breaking-change: true
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security:
  - bearerAuth: []
  - {}
...
paths:
  /v1/calendar/dates/{date}:
    get:
      tags: [Calendar]
      operationId: getCalendarDate
      summary: 查询单日
      parameters:
        - $ref: '#/components/parameters/DatePath'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Fields'
      responses:
        '200':
          description: 日期信息
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayInfo' }
              example:
                date: '2025-10-06'
                regionCode: CN
                locale: zh-CN
                localeFallback: false
                calendarSystem: GREGORIAN
                isHoliday: true
                isOfficialHoliday: true
                isWorkday: false
                isWeekend: true
                isStatutoryHoliday: true
                isAdjustedWorkday: false
                holidayNames: [国庆节, 中秋节]
                labels: [NATIONAL_DAY, MID_AUTUMN, STATUTORY]
                lunar: { year: 2025, month: 8, day: 15, isLeapMonth: false, monthName: 八月, dayName: 十五 }
                solarTerm: { index: 18, name: 寒露 }
                ganZhi: { yearName: 乙巳, heavenlyStem: 乙, earthlyBranch: 巳, zodiac: 蛇 }
                festivals: [{ code: MID_AUTUMN, name: 中秋节 }]
                sourceVersion: 2025.GOV_NOTICE
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/dates:
    get:
      tags: [Calendar]
      operationId: listCalendarDates
      summary: 查询闭区间（最多 366 天）
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Fields'
      responses:
        '200':
          description: 按日期升序的结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarResult' }
              example:
                region: CN
                locale: zh-CN
                from: '2025-10-01'
                to: '2025-10-02'
                count: 2
                items:
                  - date: '2025-10-01'
                    regionCode: CN
                    locale: zh-CN
                    localeFallback: false
                    calendarSystem: GREGORIAN
                    isHoliday: true
                    isOfficialHoliday: true
                    isWorkday: false
                    isWeekend: false
                    isStatutoryHoliday: true
                    isAdjustedWorkday: false
                    sourceVersion: 2025.GOV_NOTICE
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/dates:batch:
    post:
      tags: [Calendar]
      operationId: batchCalendarDates
      summary: 合并多个范围批量查询
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchCalendarRequest' }
            example:
              region: CN
              locale: en-US
              fields: [holidayNames, festivals]
              ranges:
                - { from: '2025-01-01', to: '2025-01-03' }
                - { from: '2025-01-03', to: '2025-01-05' }
      responses:
        '200':
          description: 合并后的范围和日期结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchCalendarResult' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/months/{year}/{month}:
    get:
      tags: [Calendar]
      operationId: getCalendarMonth
      summary: 查询月份
      parameters:
        - $ref: '#/components/parameters/YearPath'
        - $ref: '#/components/parameters/MonthPath'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Fields'
      responses:
        '200':
          description: 月份日期结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarResult' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/years/{year}:
    get:
      tags: [Calendar]
      operationId: getCalendarYear
      summary: 查询全年
      parameters:
        - $ref: '#/components/parameters/YearPath'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Fields'
      responses:
        '200':
          description: 365 或 366 天
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarResult' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/workdays/count:
    get:
      tags: [Calendar]
      operationId: countWorkdays
      summary: 统计工作日和休息日
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Region'
      responses:
        '200':
          description: 工作日统计
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkdayStats' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/holidays:
    get:
      tags: [Calendar]
      operationId: listHolidays
      summary: 查询年度节假日汇总
      parameters:
        - name: year
          in: query
          required: true
          schema: { type: integer, minimum: 2000, maximum: 2026 }
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: 节假日周期列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HolidayYearResult' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/holidays/next:
    get:
      tags: [Calendar]
      operationId: getNextHoliday
      summary: 查询下一法定节假日
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/Region'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: 从当天开始的下一法定节假日
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayInfo' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/regions:
    get:
      tags: [Calendar]
      operationId: listRegions
      summary: 列出已安装地区
      responses:
        '200':
          description: 地区、语言和数据覆盖范围
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegionCatalog' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/calendar/lunar/from-solar:
    get:
      tags: [Conversion]
      operationId: convertSolarToLunar
      summary: 公历转农历
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$', example: '2025-01-29' }
      responses:
        '200':
          description: 农历信息
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FullLunarInfo' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/solar/from-lunar:
    get:
      tags: [Conversion]
      operationId: convertLunarToSolar
      summary: 农历转公历
      parameters:
        - { name: year, in: query, required: true, schema: { type: integer, minimum: 1900, maximum: 2100 } }
        - { name: month, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 12 } }
        - { name: day, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 30 } }
        - { name: leapMonth, in: query, required: false, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: 对应公历日期
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [date]
                properties: { date: { type: string, format: date } }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/solar-terms/{year}:
    get:
      tags: [Conversion]
      operationId: getSolarTerms
      summary: 查询二十四节气
      parameters:
        - $ref: '#/components/parameters/YearPath'
      responses:
        '200':
          description: 24 个节气
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolarTermYear' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/DataNotAvailable' }
  /v1/calendar/metadata:
    get:
      tags: [Calendar]
      operationId: getCalendarMetadata
      summary: 查询契约和资产元数据
      responses:
        '200':
          description: 当前镜像元数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Metadata' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: UPSTREAM_TOKEN 配置后启用的内部令牌。
  parameters:
    DatePath:
      name: date
      in: path
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$', example: '2025-10-06' }
    YearPath:
      name: year
      in: path
      required: true
      schema: { type: integer, minimum: 1, maximum: 9999 }
    MonthPath:
      name: month
      in: path
      required: true
      schema: { type: integer, minimum: 1, maximum: 12 }
    Region:
      name: region
      in: query
      required: false
      schema: { type: string, default: CN, pattern: '^[A-Z]{2}(?:-[A-Z0-9]{1,8})*$' }
    Locale:
      name: locale
      in: query
      required: false
      schema: { type: string, enum: [zh-CN, en-US], default: zh-CN }
    Fields:
      name: fields
      in: query
      required: false
      description: 逗号分隔的可选字段白名单。
      schema: { type: string, example: 'holidayNames,lunar,festivals' }
    From:
      name: from
      in: query
      required: true
      schema: { type: string, format: date, example: '2025-01-01' }
    To:
      name: to
      in: query
      required: true
      schema: { type: string, format: date, example: '2025-01-31' }
  headers:
    ETag:
      description: 当前响应的实体标签。
      schema: { type: string }
...
    RequestId:
      description: 透传或生成的请求标识。
      schema: { type: string, maxLength: 64 }
  responses:
    InvalidArgument:
      description: 参数、日期、范围或请求体无效。
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: 'urn:api-assets:calendar:error:invalid-argument', title: 'Bad Request', status: 400, detail: 参数无效, instance: /v1/calendar/dates, code: INVALID_ARGUMENT, requestId: demo }
    DataNotAvailable:
      description: 地区或年度数据未安装，或没有后续节假日。
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unauthorized:
      description: Bearer Token 缺失或错误。
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: 'urn:api-assets:calendar:error:upstream-unauthorized', title: Unauthorized, status: 401, detail: 未授权, instance: /v1/calendar/dates, code: UPSTREAM_UNAUTHORIZED, requestId: demo }
...
  schemas:
    DayInfo:
      type: object
      additionalProperties: false
      required: [date, regionCode, locale, localeFallback, calendarSystem, isHoliday, isOfficialHoliday, isWorkday, isWeekend, isStatutoryHoliday, isAdjustedWorkday, sourceVersion]
      properties:
        date: { type: string, format: date }
        regionCode: { type: string, example: CN }
        locale: { type: string, enum: [zh-CN, en-US] }
        localeFallback: { type: boolean }
        calendarSystem: { type: string, enum: [GREGORIAN, CHINESE_LUNAR] }
        isHoliday: { type: boolean }
        isOfficialHoliday: { type: boolean }
        isWorkday: { type: boolean }
        isWeekend: { type: boolean }
        isStatutoryHoliday: { type: boolean }
        isAdjustedWorkday: { type: boolean }
        holidayNames: { type: array, items: { type: string } }
        labels: { type: array, items: { type: string } }
        lunar: { type: object, allOf: [{ $ref: '#/components/schemas/LunarDateInfo' }], nullable: true }
        solarTerm: { type: object, allOf: [{ $ref: '#/components/schemas/DaySolarTerm' }], nullable: true }
        ganZhi: { type: object, allOf: [{ $ref: '#/components/schemas/GanZhi' }], nullable: true }
        festivals: { type: array, items: { $ref: '#/components/schemas/Festival' } }
        sourceVersion: { type: string, example: 2025.GOV_NOTICE }
    CalendarResult:
      type: object
      additionalProperties: false
      required: [region, locale, from, to, count, items]
      properties:
        region: { type: string }
        locale: { type: string, enum: [zh-CN, en-US] }
        from: { type: string, format: date }
        to: { type: string, format: date }
        count: { type: integer, minimum: 1, maximum: 366 }
        items: { type: array, minItems: 1, maxItems: 366, items: { $ref: '#/components/schemas/DayInfo' } }
    BatchCalendarRequest:
      type: object
      additionalProperties: false
      required: [ranges]
      properties:
        region: { type: string, default: CN }
        locale: { type: string, enum: [zh-CN, en-US], default: zh-CN }
        fields: { type: array, minItems: 1, items: { type: string, enum: [holidayNames, labels, lunar, solarTerm, ganZhi, festivals] } }
        ranges:
          type: array
          minItems: 1
          maxItems: 32
          items: { $ref: '#/components/schemas/DateRange' }
    BatchCalendarResult:
      type: object
      additionalProperties: false
      required: [region, locale, ranges, count, items]
      properties:
        region: { type: string }
        locale: { type: string, enum: [zh-CN, en-US] }
        ranges: { type: array, items: { $ref: '#/components/schemas/DateRange' } }
        count: { type: integer, minimum: 1, maximum: 4096 }
        items: { type: array, minItems: 1, maxItems: 4096, items: { $ref: '#/components/schemas/DayInfo' } }
...
    LunarDateInfo:
      type: object
      additionalProperties: false
      required: [year, month, day, isLeapMonth, monthName, dayName]
      properties:
        year: { type: integer }
        month: { type: integer, minimum: 1, maximum: 12 }
        day: { type: integer, minimum: 1, maximum: 30 }
        isLeapMonth: { type: boolean }
        monthName: { type: string }
        dayName: { type: string }
    FullLunarInfo:
      type: object
      additionalProperties: false
      required: [date, tianGan, diZhi, ganZhiYear, shengXiao, monthName, dayName, fullName]
      properties:
        date:
          type: object
          additionalProperties: false
          required: [year, month, day, leapMonth]
          properties:
            year: { type: integer }
            month: { type: integer }
            day: { type: integer }
            leapMonth: { type: boolean }
        tianGan: { type: string }
        diZhi: { type: string }
        ganZhiYear: { type: string }
        shengXiao: { type: string }
        monthName: { type: string }
        dayName: { type: string }
        fullName: { type: string }
    GanZhi:
      type: object
      additionalProperties: false
      required: [yearName, heavenlyStem, earthlyBranch, zodiac]
      properties:
        yearName: { type: string }
        heavenlyStem: { type: string }
        earthlyBranch: { type: string }
        zodiac: { type: string }
    DaySolarTerm:
      type: object
      additionalProperties: false
      required: [index, name]
      properties:
        index: { type: integer, minimum: 0, maximum: 23 }
        name: { type: string }
    SolarTermYear:
      type: object
      additionalProperties: false
      required: [year, terms]
      properties:
        year: { type: integer }
        terms: { type: array, minItems: 24, maxItems: 24, items: { $ref: '#/components/schemas/SolarTerm' } }
...
    Festival:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
    WorkdayStats:
      type: object
      additionalProperties: false
      required: [from, to, region, calendarDays, workdays, nonWorkdays, weekendDays, statutoryHolidayDays, adjustedWorkdays]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        region: { type: string }
        calendarDays: { type: integer, minimum: 1, maximum: 366 }
        workdays: { type: integer, minimum: 0, maximum: 366 }
        nonWorkdays: { type: integer, minimum: 0, maximum: 366 }
        weekendDays: { type: integer, minimum: 0, maximum: 366 }
        statutoryHolidayDays: { type: integer, minimum: 0, maximum: 366 }
        adjustedWorkdays: { type: integer, minimum: 0, maximum: 366 }
    HolidayYearResult:
      type: object
      additionalProperties: false
      required: [year, region, locale, holidays]
      properties:
        year: { type: integer }
        region: { type: string }
        locale: { type: string, enum: [zh-CN, en-US] }
        holidays: { type: array, items: { $ref: '#/components/schemas/HolidayPeriod' } }
...
    RegionCatalog:
      type: object
      additionalProperties: false
      required: [regions]
      properties:
        regions: { type: array, items: { $ref: '#/components/schemas/RegionInfo' } }
...
    Metadata:
      type: object
      additionalProperties: false
      required: [releaseVersion, contractVersion, dataVersion, generatedAt, breakingChange, holidays, lunar, solarTerms, binaryFormats, sourceCommit]
      properties:
        releaseVersion: { type: string }
        contractVersion: { type: string, example: v1 }
        dataVersion: { type: string }
        generatedAt: { type: string, format: date-time }
        breakingChange: { type: boolean, enum: [true] }
        holidays: { type: array, items: { $ref: '#/components/schemas/HolidayCoverage' } }
        lunar: { $ref: '#/components/schemas/YearCoverage' }
        solarTerms: { $ref: '#/components/schemas/YearCoverage' }
        binaryFormats:
          type: object
          additionalProperties: false
          required: [hday, calendarData]
          properties:
            hday: { type: integer }
            calendarData: { type: integer }
        sourceCommit: { type: string }
...